+++ go/pkg/basecamp/comments.go
@@ -332,7 +332,7 @@ func commentFromGenerated(gc generated.Comment) Comment {
 
 	if gc.Creator.Id != 0 || gc.Creator.Name != "" {
 		c.Creator = &Person{
-			ID:           gc.Creator.Id,
+			ID:           int64(gc.Creator.Id),
 			Name:         gc.Creator.Name,
 			EmailAddress: gc.Creator.EmailAddress,
 			AvatarURL:    gc.Creator.AvatarUrl,
